services/resume-intel-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
pdfplumber
python-multipart
requests
